Refactor: 서버 구조 리팩토링
